Refactor: Enhance HMM ensemble training with validation and reporting
